pyproject.toml
[project]
name = "sotti"
version = "0.1.0"
description = "Local screenshot-to-code pipeline daemon."
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "watchdog>=4.0.0",
    "google-genai>=0.5.0",
    "python-dotenv>=1.2.2",
    "pillow>=12.2.0",
    "pydantic-settings>=2.13.1",
]
